refactor: scoreInputContainer 분리
